pyproject.toml
[project]
name = "zad-cli"
version = "0.1.0"
description = "CLI for ZAD (Zelfservice Applicatie Deployment)"
readme = "README.md"
license = "EUPL-1.2"
requires-python = ">=3.11"
authors = [
    { name = "Rijks ICT Gilde" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pyyaml>=6.0",
    "httpx>=0.28",
    "pydantic>=2.0",
    "websockets>=14.0",
]

[project.urls]
Repository = "https://github.com/RijksICTGilde/zad-cli"
Issues = "https://github.com/RijksICTGilde/zad-cli/issues"

[project.scripts]
zad = "zad_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0",
    "respx>=0.22",
    "ruff>=0.15.7",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.hatch.build.targets.wheel]
packages = ["src/zad_cli"]
